Move handlers from service to control-module
